backend/requirements.txt
fastapi
uvicorn
xgboost
scikit-learn
pandas
numpy
langchain
langchain-openai
python-dotenv
langchain-groq